Add Bolt Configuration Model and Table
This commit adds the bolt configuration model and the migration to create the table.

This is the first step of adding the bolt configuration option to the admin panel on solidus.

Factory for bolt_configuration has also been added and `require pry` has been added to spec_helper to help in testing.