Add Transaction Refund API Object
This commit adds the transaction refund api object.

This object inherits from the SolidusBolt::Base Service Object and allows to call the /merchant/transactions/credit api for Bolt.

This requires 3 arguments which are the `transaction_reference`, `amount` and `currency` which have to provided when the object for the class is initialized.

Bolt API DOCS reference: https://docs.bolt.com/api-bolt/#operation/refund