Add Transaction Void API Object
This commit adds the transaction void api object.

This object inherits from the SolidusBolt::Base Service Object and allows to call the /merchant/transactions/void api for Bolt.

This requires 2 arguments which are the `transaction_reference` and `credit_card_transaction_id` which have to provided when the object for the class is initialized.

Bolt API DOCS reference: https://docs.bolt.com/api-bolt/#operation/void