Add Transaction Detail API Object
This commit adds the transaction detail api object.

This object inherits from the SolidusBolt::Base Service Object and allows to call the /merchant/transactions/{transaction_reference} api for Bolt.

This requires 1 path argument which is the `transaction_reference` which has to provided when the object for the class is initialized.

Bolt API DOCS reference: https://docs.bolt.com/api-bolt/#operation/transaction-details