Swap environment_url for test column in BoltConfig
Asking users to copy-paste a url is bad UX, instead we should ask them to
decide between test and production environment and handle the related urls
with the backend logic.

This will become a greater issue down the line as the url to display iframes
is different from the one to making API calls, meaning that without this
change we would ask for even more copy-pasted urls.