Embedded Bolt Account Authorization
This commit adds the authorization process for a Bolt Account.

The commit contains a JS class that can be used to trigger the OTP authentication process for a bolt account.
The Class object requires to be initialized with an emailField, the Bolt publishable key and the Bolt domain for the required environment.

The email from the emailField is used to check if an account exists on Bolt or not. If the account exists then the authorization process is started and the response is supposed to be passed on to the backend to generate access tokens.

The space for the call to the backend has been intentionally left empty and will be populated with an API or Redirect call in the future.