Swap transaction entries for credit card entries in payment_source
When we started the project we weren't sure which kind of data we
should store in solidus_bolt_payment_sources so we decided to include
the data returned from an API call. Now it has become clear that to
properly integrate with Solidus and the rest of the app, we need to
store credit card information (which will allow us to make repeat
payments).

This commit removes original transaction-related columns and logic
in solidus_bolt_payment_sources and replaces it with credit card
columns, updating specs as needed.