Display option to create Bolt Account
Since at this stage we have all info needed to create an account, we can
provide the option to users, which we would then handle automatically in
the backend.

The selected value will be sent to the PaymentSource which will store it in an
attr_accessor as we don't need to store that info in the database.