Display option to create Bolt Account
Since at this stage we have all info needed to create an account, we can
provide the option to users, which we would then handle automatically in
the backend.

The selected value will be stored in the PaymentSource so that it can be
easily accessed at the Gateway#authorize step.