Fix Transactions::VoidService
Void API accepts transaction_reference OR transaction_id.

ref https://help.bolt.com/api-bolt/#tag/Transactions/operation/void

The documentation specifies to provide the `credit_card_transaction_id`,
but when I was testing it, I have received an error for:
`A parameter provided in the input is invalid. no transaction_id or transaction_reference entered`
Replacing the `credit_card_transaction_id` key with `transaction_id`
fixed the error.