Add addresses to Bolt
We want to upload addresses to Bolt after they have been created.
However, since in order to upload them we need the email and bolt access
token that can only be easily accessible from the controller, instead of
putting the upload address code in an AddressBook after_save callback
we are putting it in the finalize_order step of checkout flow.