Rework AuthorizeService to handle repeat payments
Until now we've been using AuthorizeService only for payments with new
credit cards. Now that we are reusing payment methods we also need to
modify the service to handle those, as the body of the request is
different (need to pass `credit_card_id` and not pass other fields).

Another point is that there have been some uncertainties with the status
of an authorized payment: we don't get the status immediately after the
call, but it takes a second or so to update. This leads to an issue
where we don't really know if the payment was authorized or failed. To
counter this we are adding extra tests to the `authorize_service_spec`
that will check the status of the payment a second after the call.