Create AddPaymentMethodService
To test the possibility of making a payment reusable, we'll need to
first create a payment via API. Since the easiest way to do so would be
via a service, this commit adds said service. As an added benefit,
future features that might also require to create a payment will be able
to use this service.