Rename Merchant ID to Division Public Id
This commit renames the previously present merchant_id field to division_public_id.

This has to be done as now Bolt has made the key public by the name of division public id which can be accessed from the Developers Dashboard on Bolt.

The column in BoltConfiguration has been renamed and all associated references have been renamed as well.