Fix Embed JS failure issue
This commit fixes the issue which caused the application to fail when loading for the first time with an empty BoltConfiguration.

The issue was that we were checking for 0 when the environment field was empty and the check for .zero? failed when used on a nil value.

This has been fixed by adding the not_null and the default value restrictions to the environment column in the table and changing the if-else clause and using a case statement instead.

The default value for the environment table is 1.