Create GetAccount endpoint
As per Bolt's request we are creating a POST endpoint to which Bolt will
pass an email param, and we'll return either 200 or 404 based on whether
there is an account created with said email. The URL which they will
have to call is `/accounts/bolt`.

Since this controller shares a lot of points in common with the existing
WebhooksController, we are also creating a BaseController from which
both will inherit.