Expire user session when access token expires
In a previous commit we setup to have the AddAddressesToBolt be called
asynchronously via a Job. However, when the first access_token expires
then new one doesn't have write access anymore so the Job fails. To
solve this we are adding a check for the expiration time and signing out
the user, to force them to get a fresh token with write access.

To improve the UX, we are moving this step to before delivery (to avoid
signing out a user right at the end of the checkout flow).