Remove the solidus_auth_devise gem from gemspec
The solidus_auth_devise gem was added into the gemspec for this project as a typo in a previous commit.

The gem is already present in the Gemfile and doesn't need to be added as a dependency in the gemspec for the project.
Hence, it is being removed in this commit.